pyproject.toml
[project]
name = "pravni-vodnik"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "grpcio>=1.76.0",
    "ipykernel>=7.1.0",
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=4.1.1",
    "langchain-groq>=1.1.1",
    "langchain-openai>=1.1.5",
    "langchain-text-splitters>=1.1.0",
    "pinecone[grpc]>=8.0.0",
    "pinecone-text>=0.11.0",
    "pymongo>=4.15.5",
    "pymupdf>=1.26.7",
    "python-dotenv>=1.2.1",
    "fastapi>=0.128.0",
    "uvicorn>=0.40.0",
    "langgraph>=1.0.5",
    "ipython>=9.8.0",
    "sse-starlette>=3.1.2",
    "numpy>=2.3.5",
]
